final-serverless/app/requirements.txt
functions-framework==3.*
Flask
sqlalchemy
gunicorn
pg8000
google-cloud-pubsub
